+++ tmail-backend/apps/distributed/pom.xml
@@ -678,6 +678,10 @@
                             </extraDirectories>
                         </configuration>
                     </plugin>
+                    <plugin>
+                        <groupId>net.alchim31.maven</groupId>
+                        <artifactId>scala-maven-plugin</artifactId>
+                    </plugin>
                 </plugins>
             </build>
         </profile>
